pyproject.toml
[project]
name = "propor-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "google-genai>=1.57.0",
    "openai>=2.14.0",
    "pymupdf>=1.26.7",
]
